Fix: main function should be matched with  (int, char**)

+++ source/app.d
@@ -34,7 +34,7 @@ void gen(ASTNode[] nodes)
     printf("  ret\n");
 }
 
-extern (C) int main(size_t argc, const(char)** argv)
+extern (C) int main(int argc, const(char)** argv)
 {
     if (argc != 2)
     {
